requirements.txt
click==8.1.7
dlib==19.24.0
face-recognition==1.3.0
face_recognition_models==0.3.0
joblib==1.4.2
numpy==1.24.2
opencv-python==4.10.0.84
Pillow==9.4.0
scikit-learn==1.5.0
scipy==1.13.1
threadpoolctl==3.5.0
